MCP Audit System/requirements.txt
fastmcp>=2.0.0
flask>=2.3.0
openai>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
markupsafe>=2.1.0
werkzeug>=2.3.0
click>=8.1.0
itsdangerous>=2.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
colorama>=0.4.6
